Cleanup NMLISTVIEW
